# src/main/resources/META-INF/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Provoly Hypervisor API
  description: CRUD Operation on repositories data
  version: 1.0.0
servers:
  - url: http://provoly-hypervisor.com
paths:
...
  /equipments:
    get:
      tags:
        - Equipment Controller
      summary: Retrieve all equipments or only equipments with entity in param
      operationId: com.provoly.event.EquipmentController_getEquipments
      parameters:
        - name: entity
          in: query
          schema:
            $ref: '#/components/schemas/EquipmentEntity'
          example: 'FAGNIERES'
      responses:
        "401":
          description: Not Authorized
        "403":
          description: Forbidden if page not specified
        "200":
          description: Returns equipment Details by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentReadDto'
              examples:
                default:
                  value: |
                    [
                      {
                        "id": "0c728960-d5fd-49ea-8fe8-f72cb9cfcccd",
                        "name": "P-1000",
                        "type": "Point Lumineux",
                        "entity": "FAGNIERES"
                      }
                    ]
      security:
        - oidc:
            - authenticated
  /equipments/id/{id}:
    get:
      tags:
        - Equipment Controller
      summary: Retrieve an equipment by id
      operationId: com.provoly.event.EquipmentController_getEquipmentsDetails
      parameters:
        - name: id
          in: path
          schema:
            $ref: '#/components/schemas/UUID'
          example: 01ffde9d-30d2-4273-b61f-c6addd8747c8
          required: true
      responses:
        "401":
          description: Not Authorized
        "403":
          description: Forbidden if page not specified
        "404":
          description: Equipment Not found
        "200":
          description: Returns equipment Details by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentReadDto'
              examples:
                default:
                  value: |
                    {
                      "id": "0c728960-d5fd-49ea-8fe8-f72cb9cfcccd",
                      "name": "P-1000",
                      "type": "Point Lumineux",
                      "entity": "FAGNIERES"
                    }
      security:
        - oidc:
            - authenticated
  /equipments/name/{name}:
    get:
      tags:
        - Equipment Controller
      summary: Retrieve an equipment by name
      operationId: com.provoly.event.EquipmentController_getEquipmentsDetails
      parameters:
        - name: name
          in: path
          schema:
            type: string
          example: P-1000
          required: true
      responses:
        "401":
          description: Not Authorized
        "403":
          description: Forbidden if page not specified
        "404":
          description: Equipment Not found
        "200":
          description: Returns equipment Details by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentReadDto'
              examples:
                default:
                  value: |
                    {
                      "id": "0c728960-d5fd-49ea-8fe8-f72cb9cfcccd",
                      "name": "P-1000",
                      "type": "Point Lumineux",
                      "entity": "FAGNIERES"
                    }
      security:
        - oidc:
            - authenticated
  /equipments/entities:
    get:
      tags:
        - Equipment Controller
      summary: Retrieve all equipment entities
      operationId: com.provoly.event.EquipmentController_getEquipmentEntitiesName
      responses:
        "401":
          description: Not Authorized
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      security:
        - oidc:
            - authenticated
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      description: Authentication
      openIdConnectUrl: https://auth/realms/provoly/.well-known/openid-configuration
  schemas:
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
      example: "ebbfbbd1-b8de-467c-8a9b-88443a49f807"
      required: true
...
    Family:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        code:
          type: string
    EquipmentEntity:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
    EquipmentReadDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        family:
          $ref: '#/components/schemas/Family'
        entity:
          $ref: '#/components/schemas/EquipmentEntity'
...